requirements.txt
pymongo
readline
ipython
pyyaml
python-dateutil
xlwt
